authEmail: move the verified-email functions to verifiedEmails.ts
The component API path changes from `components.authEmail.public.*` to
`components.authEmail.verifiedEmails.*`. This makes room for a separate
`validation.ts` module for the email validation flow.
